Apply Ruff 0.16 mechanical migration

+++ src/tet4d/engine/gameplay/pieces_shared.py
@@ -8,9 +8,8 @@ def scaled_span(
     max_cap: int | None = None,
 ) -> int:
     clamped_axis = max(1, int(axis_size))
-    scaled = max(min_size, int(round(clamped_axis * ratio)))
+    scaled = max(min_size, round(clamped_axis * ratio))
     scaled = min(scaled, clamped_axis)
     if max_cap is not None:
         scaled = min(scaled, max_cap)
     return max(1, scaled)
-
